🚚 Move standardize.ts
