+++ extensions/notion/src/utils/notion/database/property.ts
@@ -1,7 +1,7 @@
 import type { DatabaseObjectResponse } from "@notionhq/client/build/src/api-endpoints";
 
 import type { ReadablePropertyType } from "..";
-import type { Standardized } from "../../standardize";
+import type { Standardized } from "../standardize";
 
 type NotionDatabaseProperty = Extract<DatabaseObjectResponse["properties"][string], { type: ReadablePropertyType }>;
 export type DatabaseProperty = Standardized<NotionDatabaseProperty, "config">;
