+++ extensions/notion/src/utils/notion/page/property.ts
@@ -4,7 +4,7 @@ import { markdownToRichText } from "@tryfabric/martian";
 import { subMinutes } from "date-fns";
 
 import type { ReadablePropertyType } from "..";
-import type { Standardized } from "../../standardize";
+import type { Standardized } from "../standardize";
 import { getLocalTimezone } from "../global";
 
 export type PageProperty = Standardized<PageObjectResponse["properties"][string], "value">;
